[level 1] Title: 두 정수 사이의 합, Time: 0.18 ms, Memory: 31.2 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.cs
@@ -2,10 +2,10 @@ using System;
 
 public class Solution {
     public long solution(int a, int b) {
-        int max = a > b ? a : b; 
-        int min = a < b ? a : b;
+        long max = a > b ? a : b; 
+        long min = a < b ? a : b;
         
         // 등차수열의 합 : (첫째항+마지막항)*항의 개수 / 2
-        return ((long)(max+min)*(long)(max-min+1))/2; 
+        return (max+min)*(max-min+1)/2; 
     }
 }
